# Speed up dictionary merging with combiner function