@chethega 2018-06-05
Pinging @StephenVavasis because `OrderedDict` and  `SortedDict` will want to implement an optimized single-item combine-merge as well (and avoiding spurious searches in `SortedDict` is even more important than in a hash table).

As a side note, I saw that `Dict` always overwrites the old key with the new (isequal) key when updating, but this is not documented. Well, I'd imagine that it is much nicer for the garbage collector if we don't do that, because the old key (and dict) are presumably old, while the new key is much more likely to be new (young). The optimization potential especially applies for `===` non-isbits immutable keys (e.g. strings), and the semantic difference also applies for isbits; e.g. a new `NaN` key overwrites the old `NaN` key, even if they are `!==`. 

I think we consider this implementation-defined and open to change without warning? Then we probably should document this fact before release, in order to prevent people from relying on this. And maybe change it as well.