@chethega 2018-06-05
...and the PR is plain wrong, because I misunderstood the current semantics of `merge`, see [https://github.com/JuliaLang/julia/issues/27437](https://github.com/JuliaLang/julia/issues/27437).